implement day 09
This one worried me a bit, as can be seen by the fact that I wrote
tests. That said, the naive implementations I chose were fast enough;
somewhat surprisingly; I was expecting to need more algorithms
research and implementation.

This also demos my idea for carryover, which appears to work, even
though it's hardly elegant. Could be more useful in exercises which
require more calculation.

Weirdly, hyperfine claims that part 2 is substantially faster than
part 1; that's _got_ to be a testing artifact, given that part2
does a strict superset of the work of part1.

Onwards and upwards!