solve day 11
Game of life, huh? Even part 2 was pretty straightforward with the
appropriate util methods added to the `Map`. Got a confusing bug
with `project` at first, though: I'd accidentally used `filter`
instead of `take_while`, meaning that the program was exploring
to the edges of infinity searching for another point which would
lie within the bounds of the map. Once that was sorted, it was fast
and easy.