Refactor tiles: implement with DisplayWidth, not char conversions
When implementing everything by hand, it's easiest to implement
`TryFrom<char>` and `Into<char>` for simplifying reading and writing
maps. However, given the existence of `parse-display`, it's acutally
now much easier to just derive `FromStr` and `Display`. However, that
left the problem of handling the widths; it would be a bad idea to
simply naively assume that every tile is always 1 character wide!

This commit handles that case. It substantially simplifies map problems,
which is nice.