solve day 15
On the one hand, today was relatively straightforward: just implement
the iterator, and ask for the Nth. On the other hand, today is the
first day this year for which any part runs in perceptably non-
instant time, and that's not really satisfactory. I mean, I'll take
it, but I'm going to spend a little time seeing if I can speed it
up a bit.