day16: implement part 2
Used a simple naive algorithm to identify which tickets could
potentially be valid. It runs in probably O(mn), where m is the
number of fields, and n is the number of tickets. Luckily, our
input isn't huge today.

It always feels good when, the first time the solution builds,
it produces the right answer. That didn't happen today: I had to
adjust `departure_product` to use 64-bit integers instead of 32,
because on the first run, I got integer overflow in the multiplication.
Still, close enough I suppose.