solve day 17 part 2
This solution came in two parts. The first part involved the actual
solution: writing Vector4, writing PointTrait, implementing that trait
for point types in 2-4 dimensions, making ConwaySpace generic over
HighDimensionPoint: PointTrait. That was fun, and made me feel clever.

The second part came when that implementation returned the wrong
answer (though by coincidence, it was _somebody's_ right answer).
At that point, it became a matter of inventing debugging tools,
implementing the examples as tests, and tracking down what had
gone wrong.

In the end, it turned that I had made a single-character error.
The fix was this:

```diff
  pub fn inclusive_range(min: Vector4, max: Vector4) -> impl Iterator<Item = Vector4> {
      (min.x..=max.x)
          .cartesian_product(min.y..=max.y)
          .cartesian_product(min.z..=max.z)
-         .cartesian_product(min.z..=max.w)
+         .cartesian_product(min.w..=max.w)
          .map(|(((x, y), z), w)| Vector4::new(x, y, z, w))
  }
```

With that in, everything was great!

Not sure if I'm going to need this abstraction again this year, but
it's entirely possible that it'll come in handy next year. I have
the feeling that somehow we've had dimension-promotion problems
in the past, at least.