solve part 2
TBH, kind of surprised that this just worked first try.
Not sure how I feel about the extensibility today. On the one hand,
refactoring the whole AST and parser to change the operator
precedence is probably the "right" answer. On the other hand,
retaining a flat operator structure and handling the varying
precedence in code was simpler to implement today, with less
duplication of work. On the gripping hand, it would have been a
huge pain to handle the prededence like this if there were more than
two operations; in that case, I'd just need to do the proper tree
structure regardless.

Given the rarity with which I'll be designing a language whose
operator precedence changes out from under me, I think it's fine
not to have an elegant strategy in hand for dealing with that
circumstance.