LALRPOP newline parsing isn't working
Commiting this for the history, but this is the point where I think
I should just give up on LALRPOP and write my own parser. It would
in principle be possible to write a custom lexer which made this work,
but that feels like more work than writing a custom parser would be.