wip part 2
This generates _an_ answer, but not the _correct_ answer. It's too
low, and I correctly predicted why: there are cases where an earlier
rule returns successfully, but a later rule would also have returned
successfully, and produced an upstream success.

I didn't address that the first time because it implies a fair amount
of refactoring. Oh well, might as well get on it.