day 19 part 2
That was a little rough! It might not have been had I been working
on it in a more fresh condition, but the fact is that it's ten minutes
before midnight and I've had more than a little gluhwein at this point,
becuase it would unfairly take time away from my family to do it
during a vacation day.

That said, it was clear as soon as I saw part2 that there could be
alternate productions in a rule which produced differing length
"remaining" strings, so it was really a matter of just figuring out
how to effectively attack that problem. I still think that a
triply-nested loop is often a sign that refactoring is desirable,
but at this point, I don't have the energy to do that refactoring.

I'll take it.