solve day 20 part 2
The problem with manually converting an image into a list of
points is that the process is prone to hard-to-debug off-by-one
errors. This hit me here, as you can see by the 1 changed line
of code and 220 inserted lines of tests. Still, everything works
now, and that's what matters in the end.